Expose `calculateStaleAt` via `CachedMapTileMetadata.fromHttpHeaders` factory
