+++ lib/src/layer/tile_layer/tile_provider/network/caching/tile_metadata.dart
@@ -1,4 +1,5 @@
-import 'dart:io' show HttpHeaders; // web safe!
+import 'dart:io' show HttpHeaders, HttpDate; // web safe!
+import 'dart:math';
 
 import 'package:flutter_map/flutter_map.dart';
 import 'package:meta/meta.dart';
@@ -14,26 +15,77 @@ import 'package:meta/meta.dart';
 @immutable
 class CachedMapTileMetadata {
   /// Create new metadata
-  CachedMapTileMetadata({
+  const CachedMapTileMetadata({
     required this.staleAt,
     required this.lastModified,
     required this.etag,
-  })  : staleAtMilliseconds = staleAt.millisecondsSinceEpoch,
-        lastModifiedMilliseconds = lastModified?.millisecondsSinceEpoch;
+  });
 
-  /// The calculated time at which this tile becomes stale
+  /// Create new metadata based off an HTTP response's headers
+  ///
+  /// Where a response does not include enough information to calculate the
+  /// freshness age, [fallbackFreshnessAge] is used.
+  factory CachedMapTileMetadata.fromHttpHeaders(
+    Map<String, String> headers, {
+    Duration fallbackFreshnessAge = const Duration(days: 7),
+  }) {
+    // There is no guarantee that this meets the HTTP specification - however,
+    // it was designed with it in mind
+    DateTime calculateStaleAt() {
+      final addToNow = DateTime.timestamp().add;
+
+      if (headers[HttpHeaders.cacheControlHeader]?.toLowerCase()
+          case final cacheControl?) {
+        final maxAge = RegExp(r'max-age=(\d+)').firstMatch(cacheControl)?[1];
+
+        if (maxAge == null) {
+          if (headers[HttpHeaders.expiresHeader]?.toLowerCase()
+              case final expires?) {
+            return HttpDate.parse(expires);
+          }
+
+          return addToNow(fallbackFreshnessAge);
+        }
+
+        if (headers[HttpHeaders.ageHeader] case final currentAge?) {
+          return addToNow(
+            Duration(seconds: int.parse(maxAge) - int.parse(currentAge)),
+          );
+        }
+
+        final estimatedAge = max(
+          0,
+          DateTime.timestamp()
+              .difference(HttpDate.parse(headers[HttpHeaders.dateHeader]!))
+              .inSeconds,
+        );
+        return addToNow(Duration(seconds: int.parse(maxAge) - estimatedAge));
+      }
+
+      return addToNow(fallbackFreshnessAge);
+    }
+
+    final lastModified = headers[HttpHeaders.lastModifiedHeader];
+    final etag = headers[HttpHeaders.etagHeader];
+
+    return CachedMapTileMetadata(
+      staleAt: calculateStaleAt(),
+      lastModified: lastModified != null ? HttpDate.parse(lastModified) : null,
+      etag: etag,
+    );
+  }
+
+  /// The calculated time at which this tile becomes stale (UTC)
+  ///
+  /// Tile providers should use [isStale] to check whether a tile is stale,
+  /// instead of manually comparing this to the current timestamp.
+  ///
+  /// This may have been calculated based off an HTTP response's headers using
+  /// [CachedMapTileMetadata.fromHttpHeaders], or it may be custom.
   final DateTime staleAt;
 
-  /// The calculated time at which this tile becomes stale, represented in
-  /// [DateTime.millisecondsSinceEpoch]
-  final int staleAtMilliseconds;
-
-  /// If available, the value in [HttpHeaders.lastModifiedHeader]
+  /// If available, the value in [HttpHeaders.lastModifiedHeader] (UTC)
   final DateTime? lastModified;
-
-  /// If available, the value in [HttpHeaders.lastModifiedHeader], represented
-  /// in [DateTime.millisecondsSinceEpoch]
-  final int? lastModifiedMilliseconds;
 
   /// If available, the value in [HttpHeaders.etagHeader]
   final String? etag;
@@ -45,14 +97,13 @@ class CachedMapTileMetadata {
   bool get isStale => DateTime.timestamp().isAfter(staleAt);
 
   @override
-  int get hashCode =>
-      Object.hash(staleAtMilliseconds, lastModifiedMilliseconds, etag);
+  int get hashCode => Object.hash(staleAt, lastModified, etag);
 
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
       (other is CachedMapTileMetadata &&
-          staleAtMilliseconds == other.staleAtMilliseconds &&
-          lastModifiedMilliseconds == other.lastModifiedMilliseconds &&
+          staleAt == other.staleAt &&
+          lastModified == other.lastModified &&
           etag == other.etag);
 }
